feat :: api create 메서드 분리
